Debugger: Split out SymbolImporter into its own class
