+++ pcsx2/VMManager.cpp
@@ -7,7 +7,7 @@
 #include "Counters.h"
 #include "DEV9/DEV9.h"
 #include "DebugTools/DebugInterface.h"
-#include "DebugTools/SymbolGuardian.h"
+#include "DebugTools/SymbolImporter.h"
 #include "Elfheader.h"
 #include "FW.h"
 #include "GS.h"
@@ -447,8 +447,8 @@ void VMManager::Internal::CPUThreadShutdown()
 	// Ensure emulog gets flushed.
 	Log::SetFileOutputLevel(LOGLEVEL_NONE, std::string());
 
-	R3000SymbolGuardian.ShutdownWorkerThread();
-	R5900SymbolGuardian.ShutdownWorkerThread();
+	R3000SymbolImporter.ShutdownWorkerThread();
+	R5900SymbolImporter.ShutdownWorkerThread();
 }
 
 void VMManager::Internal::SetFileLogPath(std::string path)
@@ -1154,24 +1154,7 @@ void VMManager::UpdateELFInfo(std::string elf_path)
 	s_elf_text_range = elfo.GetTextRange();
 	s_elf_path = std::move(elf_path);
 
-	R5900SymbolGuardian.Reset();
-
-	// Search for a .sym file to load symbols from.
-	std::string nocash_path;
-	CDVD_SourceType source_type = CDVDsys_GetSourceType();
-	if (source_type == CDVD_SourceType::Iso)
-	{
-		std::string iso_file_path = CDVDsys_GetFile(source_type);
-
-		std::string::size_type n = iso_file_path.rfind('.');
-		if (n == std::string::npos)
-			nocash_path = iso_file_path + ".sym";
-		else
-			nocash_path = iso_file_path.substr(0, n) + ".sym";
-	}
-
-	// Load the symbols stored in the ELF file.
-	R5900SymbolGuardian.ImportElf(elfo.ReleaseData(), s_elf_path, nocash_path);
+	R5900SymbolImporter.OnElfChanged(elfo.ReleaseData(), s_elf_path);
 }
 
 void VMManager::ClearELFInfo()
